Use `CaseIterable` to auto-generate CLI `--help` entries (#8045)
The swift-argument-parser library will automatically show the available
values for options that implement the CaseIterable protocol. There are
options in SwiftPM that do either list the available values statically
in the help message, which can be stale, or don't provide a list at all,
which leaves the user to guess the value that they want. Sweep these
options, removing the static list, sometimes adding the help message
based on the swift documentation. Since the CaseIterable doesn't provide
a description of each value, leave the options alone that provide detail
for each of their available values.

The result of these changes is that the online help (and content assist)
is able to automatically display all of the available values for certain
options. Also, some of the options have a help description for the first
time.

---------

Co-authored-by: Max Desiatov <m_desiatov@apple.com>