Package plugins don't work on Windows if the PATH environment variable isn't accessed with the right case
Builds of packages using package plugins may fail on Windows with the following error:

```
error: plugin process ended by an uncaught signal: 309 <command: C:\foo\.build\plugins\cache\myplugin.exe>, <output:
>
```

309 translates to 0x135, which maps to 0xc0000135 aka STATUS_DLL_NOT_FOUND.

This is due to the following code in DefaultPluginScriptRunner.swift:

```swift
#if os(Windows)
        let pluginLibraryPath = self.toolchain.swiftPMLibrariesLocation.pluginLibraryPath.pathString
        var env = ProcessInfo.processInfo.environment
        if let Path = env["Path"] {
            env["Path"] = "\(pluginLibraryPath);\(Path)"
        } else {
            env["Path"] = pluginLibraryPath
        }
        process.environment = env
#endif
```

Environment variable names are case-insensitive on Windows. On a real Windows host, it tends to be spelled "Path". In a Windows Container (Docker), it tends to be spelled "PATH".

The code will end up clearing out the other paths from the PATH environment variable in the else case because Path != PATH.

We need to access the path here in a case-insensitive manner -- use the existing Environment abstraction for this, which handles Windows case sensitivity concerns.

Also fix another occurrence of this same pattern in the manifest loader.

Closes #8125