Use `AbsolutePath.asURL` instead of creating a new URL from the path string (#8159)
Use `AbsolutePath.asURL` instead of creating a new URL from the path
string.

### Motivation:

The original PR that added `resources`, `ignored`, and `other` files to
`BuildTarget` should have used `AbsolutePath.asURL` instead of creating
a new URL from the path string. This PR fixes that in order to maintain
a single location where absolute paths can be converted to URLs.

### Modifications:

I've replaced the `AbsolutePath` to `URL` mappings in `SourceKitLSPAPI`
with calls to `AbsolutePath.asURL`.

### Result:

This PR does not have an affect on the functionality of Swift Package
Manager.