Include the path to the platform's PrivateFrameworks directory in DYLD_FRAMEWORK_PATH when launching test runners (#8199)
This modifies the logic which sets the value of the
`DYLD_FRAMEWORK_PATH` environment variable when launching test runners
on Darwin when Xcode is installed: currently it includes the path to the
platform's developer "Frameworks" directory, and this change adds the
sibling "PrivateFrameworks" directory as well.

### Motivation:

XCTest includes some private support frameworks, and these frameworks
must be aligned with the public framework for testing to work correctly.
This ensures that if the path to the (public) frameworks directory is
specified, the aligned private frameworks directory is also included.

### Modifications:

- Add `$PLATFORM/Developer/Library/PrivateFrameworks` to the
`DYLD_FRAMEWORK_PATH` env var.

Resolves rdar://141564191